update and end tp

+++ src/Controller/WeaponController.php
@@ -6,6 +6,8 @@ use App\Entity\Weapon;
 use App\Entity\WeaponType as WeaponTypeEntity;
 use App\Form\WeaponType;
 use App\Repository\WeaponRepository;
+use App\Security\Voter\AppAccess;
+use App\Service\Weapon\Load;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -35,7 +37,7 @@ class WeaponController extends AbstractController
     public function new(Request $request, WeaponTypeEntity $WeaponType = null): Response
     {
         $weapon = new Weapon();
-        $form = $this->createForm(WeaponType::class, $weapon);
+        $form = $this->createForm(WeaponType::class, $weapon, ['weapon_type' => $WeaponType]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -43,6 +45,9 @@ class WeaponController extends AbstractController
             $entityManager->persist($weapon);
             $entityManager->flush();
 
+            if($WeaponType !== null){
+                return $this->redirectToRoute('user_profile');
+            }
             return $this->redirectToRoute('weapon_index');
         }
 
@@ -54,26 +59,42 @@ class WeaponController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="weapon_show", methods={"GET"})
+     * @Route("/{id}", name="weapon_show", methods={"GET"}, requirements={"id"="\d+"})
      */
     public function show(Weapon $weapon): Response
     {
+        if($this->isGranted(AppAccess::WEAPON_SHOW, $weapon) === false){
+            $this->addFlash('error', 'you cannot access to this object !');
+            return $this->redirectToRoute('weapon_index');
+        }
+
         return $this->render('weapon/show.html.twig', [
             'weapon' => $weapon,
         ]);
     }
 
     /**
-     * @Route("/{id}/edit", name="weapon_edit", methods={"GET","POST"})
+     * @Route("/{id}/edit/{profile}", name="weapon_edit", methods={"GET","POST"}, defaults={"profile":null})
      */
-    public function edit(Request $request, Weapon $weapon): Response
+    public function edit(Request $request, Weapon $weapon, string $profile = null): Response
     {
-        $form = $this->createForm(WeaponType::class, $weapon);
+
+        $this->denyAccessUnlessGranted(AppAccess::WEAPON_EDIT, $weapon);
+
+        $options = [];
+        if($profile !== null){
+            $options = ['ammunition' => true];
+        }
+
+        $form = $this->createForm(WeaponType::class, $weapon, $options);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
 
+            if($profile !== null){
+                return $this->redirectToRoute('user_profile');
+            }
             return $this->redirectToRoute('weapon_index');
         }
 
@@ -88,6 +109,8 @@ class WeaponController extends AbstractController
      */
     public function delete(Request $request, Weapon $weapon): Response
     {
+        $this->denyAccessUnlessGranted(AppAccess::WEAPON_DELETE, $weapon);
+
         if ($this->isCsrfTokenValid('delete'.$weapon->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($weapon);
@@ -95,5 +118,18 @@ class WeaponController extends AbstractController
         }
 
         return $this->redirectToRoute('weapon_index');
+    }
+
+
+    /**
+     * @Route("/load/{id}", name="weapon_load", methods={"GET"}, requirements={"id"="\d+"})
+     */
+    public function load(Weapon $weapon, Load $load): Response
+    {
+        //$this->addFlash('success', 'Crick Crick');
+
+        $load->load($weapon);
+
+        return $this->redirectToRoute('user_profile');
     }
 }
